fix(static pages) folder structure update
