+++ packages/storefront/pages/static/service.tsx
@@ -1,16 +1,17 @@
 import type { NextPage } from "next";
 import Image from "next/image";
 
-import PageTitle from "@/components/pageTitle";
-import ThreeStoriesCard from "@/components/threeStoriesCard";
-import Container from "@/components/container";
+import PageTitle from "@/components/common/components/pageTitle";
+import ThreeStoriesCard from "@/components/static/threeStoriesCard";
+import Container from "@/components/common/components/container";
 
-import { storiesBody } from "../utils/types";
+import { storiesBody } from "../../utils/types";
 
 import styles from "@/styles/Service.module.css";
 
 const threeStoriesList: storiesBody[] = [
   {
+    id: "507f1f77bcf86cd799439011",
     title: "Our Mission",
     image:
       "https://cdn.shopify.com/s/files/1/0359/6350/2651/files/about2_large.jpg?v=1588135255",
@@ -18,6 +19,7 @@ const threeStoriesList: storiesBody[] = [
       "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto",
   },
   {
+    id: "507f1f77bcf86cd799439012",
     title: "History Of Us",
     image:
       "https://cdn.shopify.com/s/files/1/0359/6350/2651/files/about3_large.jpg?v=1588135320",
@@ -25,6 +27,7 @@ const threeStoriesList: storiesBody[] = [
       "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto",
   },
   {
+    id: "507f1f77bcf86cd799439013",
     title: "What Do We Do?",
     image:
       "https://cdn.shopify.com/s/files/1/0359/6350/2651/files/about4_large.jpg?v=1588135356",
