+++ packages/storefront/utils/types.ts
@@ -1,15 +1,18 @@
 export interface accordionBody {
+    id:string,
     title: string;
     body:string
   }
 
   export interface storiesBody {
+    id:string,
     image: string;
     title: string;
     description:string
   }
 
   export interface CarouselBody {
+    id:string;
     title: string;
     body:string;
     image: string;
